resolve reviewers comments
